Fix registration failing with PostgreSQL 9.3
Updating a row causes sys_object_bu to fire which updates sys_objects which
causes sys_object_au to fire which again updates the original table.
PostgreSQL 9.3 detects that this causes the row to change before it is updated
by the original query. Previously this was ignored.
Fix by detecting the recursion.